Removed all redundant functionalities
