+++ src/superphot_plus/supernova_class.py
@@ -11,7 +11,6 @@ class SupernovaClass(str, Enum):
     SUPERNOVA_II = "SN II"
     SUPERNOVA_IIN = "SN IIn"
     SUPERLUMINOUS_SUPERNOVA_I = "SLSN-I"
-    # SUPERLUMINOUS_SUPERNOVA_II = "SLSN-II"
 
     @classmethod
     def get_type_maps(cls, allowed_types=None):
