+++ tests/superphot_plus/samplers/test_dynesty_sampler.py
@@ -4,25 +4,24 @@ from superphot_plus.samplers.dynesty_sampler import DynestySampler
 
 
 def test_dynesty_single_file(
-    single_ztf_lightcurve_object,
+    test_ztf_photometry,
     ztf_priors,
-    single_ztf_lightcurve_fit
+    test_sampler_result
 ):
     """Just test that we generated a new file with fits"""
-    sampler = DynestySampler()
-    posterior_samples = sampler.run_single_curve(
-        single_ztf_lightcurve_object, priors=ztf_priors, rstate=np.random.default_rng(9876)
+    sampler = DynestySampler(
+        priors=ztf_priors, random_state=np.random.default_rng(9876)
     )
-
-    sample_mean = posterior_samples.sample_mean()
-    assert len(sample_mean) == 15
+    sampler.fit_photometry(test_ztf_photometry)
+    sample_mean = np.mean(sampler.result.fit_parameters, axis=0)
+    assert len(sample_mean) == 14
 
     # Check that the same means the same order of magnitude (within 50% relative value).
     # Despite setting the the random seed, we still need to account (so far) unexplained
     # additional variations.
-    expected = single_ztf_lightcurve_fit
-    assert len(expected) == len(sample_mean)
-    assert np.all(np.isclose(sample_mean, expected, rtol=0.5, atol=0.2))
+    expected_mean = np.mean(test_sampler_result.fit_parameters, axis=0)
+    assert len(expected_mean) == len(sample_mean)
+    assert np.all(np.isclose(sample_mean, expected_mean, rtol=0.5, atol=0.2))
 
     ## could be between ~600 and ~800, and can vary based on hardware.
-    assert 600 <= len(posterior_samples.samples) <= 1000
+    assert 600 <= len(sampler.result.fit_parameters) <= 1000
